+++ DatadogInternal/Tests/NetworkInstrumentation/BaggageHeaderMergerTests.swift
@@ -5,8 +5,7 @@
  */
 
 import XCTest
-import TestUtilities
-@testable import DatadogRUM
+@testable import DatadogInternal
 
 class BaggageHeaderMergerTests: XCTestCase {
     // MARK: - Basic Functionality Tests